Move map into a new package
